SY-4647: Run staged-data migrations after all services open
